requirements-dev.txt
# SPDX-FileCopyrightText: 2026 The AIDA Core Authors
# SPDX-License-Identifier: MPL-2.0

# AIDA Core Plugin Development Dependencies
#
# Installed into ~/.aida/venv/ alongside runtime deps.
# Install with: make install

# Testing
pytest>=8.0
pytest-cov>=4.0

# Linting
ruff>=0.4.0
yamllint>=1.35
reuse>=4.0
